+++ tests/lib/Acceptance/Relations/MorphOne/UpdateTest.php
@@ -47,7 +47,7 @@ class UpdateTest extends TestCase
         ]);
     }
 
-    public function testImageToNull(): void
+    public function testImageToNullKeepsImage(): void
     {
         $post = Post::factory()->create();
         $image = Image::factory()->create([
@@ -69,7 +69,29 @@ class UpdateTest extends TestCase
         ]);
     }
 
-    public function testImageToImage(): void
+    public function testImageToNullDeletesImage(): void
+    {
+        $this->schema->relationship('image')->forceDeleteDetachedModel();
+
+        $post = Post::factory()->create();
+        $image = Image::factory()->create([
+            'imageable_id' => $post->getKey(),
+            'imageable_type' => Post::class,
+        ]);
+
+        $this->repository->update($post)->store([
+            'image' => null,
+        ]);
+
+        $this->assertTrue($post->relationLoaded('image'));
+        $this->assertNull($post->getRelation('image'));
+
+        $this->assertDatabaseMissing('images', [
+            $image->getKeyName() => $image->getKey(),
+        ]);
+    }
+
+    public function testImageToImageKeepsOriginalImage(): void
     {
         $image1 = Image::factory()
             ->for(Post::factory(), 'imageable')
@@ -101,6 +123,41 @@ class UpdateTest extends TestCase
             'id' => $image1->getKey(),
             'imageable_id' => null,
             'imageable_type' => null,
+        ]);
+    }
+
+    public function testImageToImageDeletesOriginalImage(): void
+    {
+        $this->schema->relationship('image')->forceDeleteDetachedModel();
+
+        $image1 = Image::factory()
+            ->for(Post::factory(), 'imageable')
+            ->create();
+
+        $image2 = Image::factory()
+            ->for(Post::factory(), 'imageable')
+            ->create();
+
+        $post = $image1->imageable;
+
+        $this->repository->update($post)->store([
+            'image' => [
+                'type' => 'images',
+                'id' => (string) $image2->getRouteKey(),
+            ],
+        ]);
+
+        $this->assertTrue($post->relationLoaded('image'));
+        $this->assertTrue($image2->is($post->getRelation('image')));
+
+        $this->assertDatabaseHas('images', [
+            'id' => $image2->getKey(),
+            'imageable_id' => $post->getKey(),
+            'imageable_type' => Post::class,
+        ]);
+
+        $this->assertDatabaseMissing('images', [
+            $image1->getKeyName() => $image1->getKey(),
         ]);
     }